Wrote individual calibration step scripts
The calibration pipeline has now been split into discrete logical steps,
each of which has a dedicated script that can be called by CASA/SLURM.
These scripts are yet to be tested - Once Jordan's slurm builder scripts
are ready, everything can be tested together.

The directories now also use a nested structure, in order to have better
global namespace behaviour. That is, the previous commit organised the
scripts in the processMeerKAT directory (i.e.,
processMeerKAT/<scripts>), and hence any particular module (such as
config parsers) would have to be imported as

   import config_parsers

from a python script/session. This is obviously not optimal, as a module
name within processMeerKAT could conflict with other python packages
installed on the system leading to unpredictable behaviour.

With the new nested structure, all the scripts are stored in
processMeerKAT/processMeerkat/<scripts>. And the packages can now be
imported as

   import processMeerKAT.config_parsers

which is obviously more desireable behaviour.