Messy commit
Contains a bunch of fixes to the pipeline in order to get it working.
A lot of the fixes are hack-y, and several paths have been hardcoded in
the python scripts. Once it has been cleaned up and tested, it will be
merged in to the master branch.

The calibration pipeline is functioning, which was the goal. The code
has to be cleaned up and made more flexible in order to be more
generally useable.