Implemented input validation, refant calculation
Added a function that validates the type of the input from the config
file - particularly useful for stripping the `/` at the end of the
visibility name.

Also implemented some logic to calculate the reference antenna from the
visibilities (cal_scripts/calc_refant.py).

Added the specification of bad frequency ranges to the config file,
which was earlier not implemented in the scripts.

Some minor changes to the default config file - Number of channels to
average after calibration is now 1 (was 2).