Updated arguments
Updated arguments in default_config.txt and how they're read by processMeerKAT.py. Option 'scripts' is now a list of tuples with indices corresponding to the script to run, whether it's threadsafe, and which container to use to run it ('' to use default container). Changed --nosubmit to --submit. Pipeline now writes a .config.tmp file that is read in, so user can't break it as easily. Added option for mpi wrapper to use.