Fixed bug with parallel_cal.py and default_config
There was a typo in default_config.txt, in the `standard` keyword, which
has been fixed.

cal_scripts/parallel_cal.py was calling the arguments to
config_parser.overwrite_config in the wrong order, and so was causing
the script to crash. It now calls the arguments with the correct keywords (order-agnostic).