Bugfix for ant check in partition.py
The reference antenna check in `get_fields.py` depends on the CASA `msmd`
toolkit, which CASA does not make available to imported modules. So if a script
called `script1.py` is executed within the CASA environment, it has the toolkit
available to it. However if `script1.py` imports `script2.py`, `script2.py` will
not see the toolkit.

The workaround is to import the `casac` module, and initialise the tools
required through that module. For example the `msmd` module would be initialised
as

```
import casac

msmd = casac.casac.msmetadata()
```

and then used as per usual.