Added plotting functionality
Wrote plot_solutions.py to plot bandpass and phase calibration solutions using plotms, which is now included as the last script in the pipeline. This assumes xvfb-run is installed in the container running the script, which is true for /data/exp_soft/pipelines/casa-stable-5.4.1-31-dev.simg, but not the default container, which needs to be updated.

I also committed fastplot.py for anyone who might want to use that, but it is not included in the pipeline.

Fixed a bug where container and mpi wrapper weren't updated when passed in via command line during building config file. Also fixed a bug where scripts in bash path weren't being found.

Output from initial CASA call to extract field IDs is now supressed.