Added script to run a quick tclean at the end
At the moment it is in the cal_scripts/ directory, but it should ideally be
moved out since it is not a "calibration" script. It makes a quick image of the
phase cal and target (2000 iterations each) so the user can eyeball the quality
of calibration, along with inspecting the cal tables etc.