Updates to SLURM parameters and field IDs
Updated memory limit (per node) to 244 GB (250000 MB), with all memory arguments now in GB, and the default memory (per node) as this limit. Updated default number of nodes to 10 (i.e. 10 nodes x 8 tasks per node = 80 threads) to suit a typical number of scans. By default, if there are too few or too many threads compared to scans, a warning will be displayed and the number of nodes and tasks per node will be overwritten in the config file, to the lowest number of nodes (i.e. multiple of tasks per node) that is above a limit defined by the number of scans + 1 + 10%. Or if the maximum node count is reached, the number of tasks per node will increase by one until either of both limits is reached (scans + 1 + 10% or maximum number of tasks per node).

Added option [-l --local] to allow for the initial CASA call to extract fields to be made without srun (e.g. on login node, or another fat node). If srun is used, a small job is deliberately requested for efficiency, which uses 1 task on 1 node, with 4 GB memory and a 5 minute time limit.

config_parser.py and and get_fields.py now use logging, rather than print statements. get_fields.py now only opens the MS once (with msmd), rather than several times, which was costly due to the internal consistency checks that were performed upon opening.

Updated how the field IDs are extracted. Only the first field ID for the total flux calibrator is extracted, and all others are appended to the phase calibrator list (if not already present). Warnings are displayed when there are multiple IDs. If intent 'CALIBRATE_PHASE' isn't present, 'CALIBRATE_AMPLI' is used.