Modified get_fields.py
Modified get_fields.py to use the field with the maximum number of scans in the
case of multiple fields with the CALIBRATE_PHASE intent. This is to solve the
issue of passing multiple phase calibrators in to gaincal during cross
calibration. The Xf+QU gain mode also needs a fractional source model to be
passed in, and at the moment gaincal cannot accept more than one source model at
a time.