Calculate refant after patition + update tclean
Moved calculation of reference antenna to new script 'calc_refant.py', to be run after partition, so we don't search through the entire MS when calculating the refant, but rather a potentially smaller MMS.

Added real,imag plot for phase calibrator to `plot_solutions.py'.

Updated 'quick_tclean.py' to only use 1 TT when bandwidth < 100 MHz, to image all calibrated fields with imsize of 512, and to include the field name (not ID) in the output image names. This script is now called with 4 cores per task.