Updated filenames & variable names & new validation checks
I've updated the filenames and variable names, both for the config file, and the command line arguments, to what I suggest we freeze them on. I've kept in mind we want to see which script is running in the SLURM queue, so have removed unnecessary prefixes like 'cal' and 'run'. I've also updated some of the default values and comments.

Memory limit updated to 236 GB. A number of additional validation checks have been added, including verifying the path to the MS in the run stage (as well as build stage), the value of [-P --plane], and during the run stage, that all sections are present, and all expected variables within those sections are present.

Added functionality for config section comments, but as this doesn't seem very useful right now, it currently does nothing.