Updated jobs for plotting and tclean, fluxscale output, fixed keepmms bugs and updated field extraction
Only call xvfb-run when 'plot' is in the name of the script being run, to reduce chance of xvfb errors causing timeouts on jobs such as flagging. If keepmms=False, tclean will be forced to call srun with a single task and node. Increased number of CPUs per task to 8 for tclean.

Fluxscale now writes 'fluxscale_xx_yy.txt' and 'fluxscale_xy_yx.txt', and displays extra information in the logs.

Fixed bugs where split was naming output files 'mms', and booking and tclean assumed extension 'mms'.

Added functionality to 'plot_solutions.py', which plots amplitude as a function of frequency, to eyeball RFI in each of the split datasets. Changed all 'chan' to 'freq'.

In the case where intent "CALIBRATE_FLUX" is missing, processMeerKAT.py is setup to report KeyError and terminate, and the config file does not write properly. This is now updated to display logger error, and then user is required to input their fields manually. If they miss the error, the run step will detect the missing fields and will terminate.