Don't process flagged SPWs, concat, relative path support & improved build step
When nspw > 1, completely flagged SPWs are not processed at all, but removed from the config (only works for MHz format right now, but when MHz not used, warning is output). nspw forced to be an integer.

Added concat.py script to aux_scripts directory, which currently only concatenates MSs, and not MMSs, as well as produces a cube of all the sucessfully imaged SPWs when nspw > 1. Pipeline now looks for scripts in this and cal_scripts directory, and one directory up, during the run step. Tested this successfully. This isn't automatically done currently, but can be launched with a separate config that just includes concat.py.

When nspw > 1 and a relative path is input, '../' is appended as a prefix to the MS name in each SPW directory.

Build step now calls casacore inside container, rather than launching CASA. And qos is now included to increase chance of being assigned resources for the build step. Specific output is supressed to make this nicer to the user. The build step now informs the users of the reference antenna selection, and exluded nodes requested for exclusion via command line.

quick_tclean now always run in parallel, since this works even for an MS. It does not attempt to detect when mpicasa is not being used, but I don't anticipate this as a problem. Bug fixed where bandwidth variable was a list when nspw > 1, which is now summed.

Updated default container to CASA 5.6.2 and pipeline version to 1.1.

Please test and report bugs. More enhancements coming soon!