Bugfix: setjy.py incorrect flux scaling
The modifications to add full Stokes models of either 3C138 or 3C286 in
setjy had a bug where it would set the full Stokes model for _all_ the
fields in the MS. This is now fixed.

Also a small bugfix for calc_refant.py to account for antennas that
exist in the metadata but have no correlations.