New ant stats script, new selfcal bookkeeping, reordered config, smaller changes
Added new auxilary script 'show_ant_stats.py', which produces two plots of the antenna flagging statistics, and which reference antenna was chosen.

Added get_selfcal_params() to bookkeeping.py, which supports using a list for all parameters other than nloops and restart_no, and checks that the length of the lists is nloops + 1, or for solint, is nloops. It also supports a list of lists for imsize and multiscale.

Reordered the default config file, so crosscal and selfcal sections come last.

Added a check of the selfcal parameters during the run step.

Other smaller changes. All of these changes have been tested for both the selfcal scripts and the cross-cal scripts.