Calculate reference antenna when nspw>1, pre-averaging, updated vis during split, and more
Added functionality to calculate the reference antenna before running partition, for the case when nspw > 1, where the reference antenna must be the same over all SPWs. This will select the antenna with the least amount of flags within the raw data, and overwrite the reference antenna in the config file of all SPW directories. When nspw=1, this step will occur after the first round of flagging, to inform a better selection of reference antenna. Updated the default config to calcrefant=True.

Updated config and partition.py script to allow for frequency averaging before calibration. This averages n frequency channels together, while still using the 'specavg' parameter to average further frequency channels together after calibration. The mstransform call within partition.py will now set nthreads to the number of polarisations, if that value is enabled by the SLURM sbatch --cpus-per-task argument, otherwise 1.

The final split will now overwrite the vis argument in the config with the first target field, ready to be imaged by selfcal. It writes the previous vis to 'crosscal_vis', which is called by quick_tclean.py.

Other minor bugfixes and enhancements.