Bugfixes
Fixed bug in selfcal bookkeeping to do with the length of the gaincal parameters. Fixed typo in error msg. Added has_section() function to config_parser.py. Run step now checks if selfcal section is present using this function, upon which it validates the section, rather than relying on the user passing in [-2 --do2GC] again.