Support for plotcal_spw, parse SPW, and other improvements
Added 'plotcal_spw.py' to postcal scripts, and edited this script to write to "plots" directory, in pdf format only. Commented out plotting of solutions in 'plot_solutions.py', leaving only the plotting of corrected data. Added new checks to ensure spw from config is parsed properly, for this plotting script and 'concat.py’, allowing for alternative formats such as GHz, kHz, or channel. When nspw=1, concat will not run, and the first (duplicate) calc_refant.py script will be dropped, in preference for the one that follows the first round of flagging. General improvement to error checking and output msgs.