Adding science imaging section / script for final tclean
Added [image] section and science_image.py script for performing final tclean. This includes the same tclean arguments as in the [selfcal] section, as well as multiscale, restoringbeam and specmode, but is only exposed in the config if option [-I --science_image] is used. An RMS map (for a S/N-based threshold), and mask are included in this section and written to it if it exists during selfcal.