Bugfixes for plotcal_spw
The XY phase plots had the name of the secondary cal, when it should
have had the name of the pol cal. The plot still had the right data, it
just printed the wrong name.

Switched the default extension of the plots from PDF to PNG for easier
vieweing through different tools.

If the caltable didn't exist, the routine would print a warning and try
and continue with the plot and crash at the end when it couldn't read
the data. Now it prints the warning and skips the plotting to continue
with the rest.