New functionality for outliers, skipping / discarding selfcal, and seperate blind tclean parameters
New functionality has been added that enables building a tclean 'outlierfile', which enables imaging a smaller region and still deconvolving bright off-axis sources. This functionality takes all sources above a user-set threshold from the first blind tclean. Only sources sitting outside imaged region are included as outliers, so the initial image needs to be large (and optionally poor resolution, for speed). The positions and masks of the outlier sources are updated during each loop, and therefore, this initial building of the outlierfile will soon be replaced by a sky model from surveys like SUMSS. Currently, the default is outlier_threshold=0.0, meaning all of this functionality is disabled, as it is still being tested.

New functionality has been added to skip the solving and applying steps during a given loop (by setting calmode = '' for a given list's index), and for discarding the solutions for a fixed number of loops during subsequent solving and applying ('discard_nloops').

New functionality has been added where the imaging parameters of the initial blind tclean correspond to index 0 of the imaging parameter lists, and each subsequent image gets its own unique set of parameters. This is optional, as a single set of imaging parameters can be used for every loop by providing single values (instead of lists).

Lastly, the book keeping for the selfcal scripts has been improved.