Added PB correction to science_image.py
science_image.py now produces an additional PB corrected image (using
katbeam) at the end of imaging. The cutoff for the PB can be specified
by the pbthreshold parameter in the imaging section of the config file.

The default container has been updated to use the new CASA 6.5 release,
the container also contains katbeam as a dependency.